requirements.txt
django==3.0.7
pix2tex
pillow